requirements.dev.txt
-r requirements.dev.txt
asgiref==3.4.1
python-decouple==3.4
pytz==2021.1
sqlparse==0.4.1
typing-extensions==3.10.0.0
